Small Win64 fix for fscan_f and props paths for potential x64 and arm64 builds on Windows (though only Win32 necessary for actual release at this point)

+++ src/djfile.h
@@ -1,4 +1,4 @@
-//Copyright (C) 1995-2022 David Joffe / Dave Gnukem project
+//Copyright (C) 1995-2023 David Joffe / Dave Gnukem project
 //
 //dj2022-11 refactoring some file stuff into new djfile.h/cpp and maybe adding some more file- and path-related helpers and maybe some
 
@@ -56,9 +56,12 @@ extern bool djEnsureFolderTreeExists(const char* szPath);
 #ifdef djHAVE_SAFER_FUNCTIONS
 	#define dj_fscanf fscanf_s
 	#define dj_sscanf sscanf_s
+	// From https://learn.microsoft.com/en-us/cpp/c-runtime-library/reference/fscanf-s-fscanf-s-l-fwscanf-s-fwscanf-s-l?view=msvc-170 :
+	// "the more secure functions require the size in characters of each c, C, s, S, and [ type field to be passed as an argument immediately following the variable"
+	// but NB note also we get: warning C4477: 'fscanf_s' : format string '%s' requires an argument of type 'unsigned int', but variadic argument 2 has type 'size_t'
 	#define	dj_fscanf_intline(pStream, nInt)  fscanf_s((pStream), "%d\n", &nInt)
 	#define	dj_fscanf_int(pStream, nInt)      fscanf_s((pStream), "%d", &nInt)
-	#define	dj_fscanf_line(pStream, szBuf)    fscanf_s((pStream), "%s\n", szBuf, sizeof(szBuf))
+	#define	dj_fscanf_line(pStream, szBuf)    fscanf_s((pStream), "%s\n", szBuf, (unsigned int)sizeof(szBuf))
 #else
 	#define dj_fscanf fscanf
 	#define dj_sscanf sscanf
